Updated for DTOs, Domain, Repository
